CRAYSAT-945: Add a field for NodeAccelLocationInfo.Name to the NodeAccel
Add a field to the output of `sat hwinv --list-node-accels` that reports
the value of the "Name" field from the "NodeAccelLocationInfo" field for
objects of type NodeAccel in HSM. This field was found to be useful for
differentiating between actual GPUs and items that are not GPUs (items
like "NV Switch" and "NVIDIA HGX_3") on the Proliant DL385 UANs with
NVIDIA A100 GPUs. The field will be titled "Location Name" which
indicates it is the name from the location info in HSM.

Once these objects that are not really GPUs are handled better in
Redfish and/or HSM, we can remove this field because it doesn't provide
much info for other hardware like GPUs in Grizzly Peak blades, for
instance.

Test Description:
Tested by running `sat hwinv --list-node-accels` on baldar where there
is a DL385 UAN with NVIDIA A100 GPUs. Also ran `sat hwinv --list-nodes`
and `sat hwinv --summarize-nodes` for good measure since the NodeAccel
objects get associated with the nodes for reporting counts of NodeAccel
objects at the node level. These objects are still included in those
counts because we don't want to special-case and hardcode certain values
in sat.

New unit tests pass. pycodestyle passes.