CRAYSAT-931: Remove dependency on slingshot-tools RPM
Remove the dependency of the cray-sat RPM on the slingshot-tools RPM.
This was originally added because `sat cablecheck` called
`check-hsn-cables` which was delivered by that RPM. However, `sat
cablecheck` was deprecated and then removed in favor of functionality
delivered by the Slingshot product itself, specifically the Slingshot
Topology Tool (STT).

Note that we do not currently support installation of this RPM, but this
clean up of an unused dependency does not hurt anything, and it will
make it easier to support RPM installation in the future if we need to.

Test Description:
Will download RPM built by Jenkins and query it to confirm it does not
require a slingshot-tools package anymore.