CRAYSAT-721: Add bmccreds subcommand
The `sat bmccreds` subcommand provides a simple interface for
setting BMC passwords.

The command uses HSM to check the supplied xnames match the
desired types, or if no xnames are given to get a full list
of controllers. Optionally, disabled controllers may be addressed.

The command can take a password on the command line, prompt for
a password, or generate random passwords. When generating random
passwords, the `--pw-domain` argument controls whether some BMCs
may get the same randomly-generated password.

This commit enhances the `get_username_and_password_interactively`
utility function to optionally prompt for password confirmation,
which is desirable when setting a new password. This commit also
adds unit tests for this utility.

Passwords are validated against fixed criteria such as length
and character set, and are also checked for a minimum strength
using the `passwordmeter` library. This commit updates the
requirements files to include `passwordmeter` and its dependencies.

To set the BMC credentials, the command uses the SCSD API.

This commit adds get_chassis() and get_cabinet() functions to the
XName class in xname.py which are used for grouping xnames.

Add change log entry and man page. This commit also slightly modifies
the developer documentation for adding a new man page.

Test Description:
* I installed this version of `sat` in a Virtualenv on a Shasta system
  and ran `sat bmccreds` several ways:
  * `sat bmccreds` without arguments to test setting all BMCs to
    a known password. Tested with a 'weak' password to verify that
    password strength was being checked.
  * `sat bmccreds` with a 'weak' password and `--no-pw-strength-check`
    to verify that the strength check could be bypassed.
  * `sat bmccreds --random-password` with no arguments to test setting
    all BMC passwords to random values. To test this, I used the
    vault to read the passwords that had been set using SCSD.
  * Repeated the above with `--domain cabinet` and `--domain bmc`
    to verify that randomly-generated passwords could be grouped by
    cabinet or BMC.
  * Ran `sat bmccreds` with `--no-hsm-check`, and used
    `--loglevel=debug` to verify that HSM was not being queried.
  * Ran with `--include-disabled` on a system that had two disabled BMCs.
  * Ran with `--include-failed-discovery` on a system that had a BMC
    on which discovery had failed.
  * Ran with `--disruptive` to check that it did not prompt.

* I compiled and viewed the `sat-bmccreds` man page, and the main `sat`
  man page.
* I viewed the output of `sat -h` to verify the new command was present
  in the help text.

Test Description:
I re-ran some of the tests from the previous commit:
  * sat bmccreds --random-password --pw-domain bmc
  * sat bmccreds --password **password** --no-pw-strength-check
I checked the vault after setting the passwords to make
sure they were the expected (or random) values.