CRAYSAT-932: Fix help text for `sat status --type`
This change adds help text to `sat status --type` which lists all
available types which may be queried. This is a similar solution as the
one used in `sat sensors`, so it may be prudent to factor out the common
functionality between the two.

Test Description:
Run `sat status --help` locally and confirm the help text for `--type`
is correct.