CRAYSAT-964: Add SubRole to SAT status command
This commit adds a new column to the output of `sat status` describing
the HSM subrole (e.g. Master, Worker, Storage for Management nodes and
UAN for Application nodes).

This commit also changes the constant values of component fields and
table headers from two parallel tuples to an OrderedDict, mapping the
API keys to their corresponding column headers.

This commit introduces a special case for how missing values are handled
specifically for SubRole. Because Compute nodes are expected to not have
a SubRole, it looks slightly better to use the string 'None' rather than
MISSING. This behavior only applies to Compute nodes which have no
subrole. Added unit tests for this behavior.

Finally, this commit changes the sat-status man page and adds a
CHANGELOG entry. A couple minor formatting issues are corrected in
the `sat-status` man page.

Test Description:
* I ran `sat status` several ways to test a few different cases:
** `sat status` without any arguments
** `sat status --format YAML`
** `sat status --sort-by subrole`
* I compiled and viewed the man page.