CRAYSAT-824: Improve firmware error reporting
This change makes it so that if FAS returns a target without an xname
but with errors, that is logged properly and the admin is alerted that
something is going on with FAS. In addition, if a valid target is
returned without an xname, it will be logged in the report with simply
the xname missing.

Test Description:
Run unit tests. Run SAT on baldar and confirm that if the error is
replicated as in the ticket, it instead prints an improved warning
message.