CRAYSAT-957: Show empty and missing columns if a field was specified
This change slightly modifies the functionality of the --*-fields
options in `sat hwinv`. If a field name is specified but all its members
are EMPTY or MISSING, the column of that field will now be displayed
regardless of the --show-empty/missing command line options or the
show_empty/missing value in the configuration file.

Test Description:
Run on baldar to confirm new functionality works as intended.