CRAYSAT-957: Do not remove filtered columns if missing or empty
This change modifies the behavior of commands which support --filter by
preventing columns that are touched by a filter from being removed if
their rows are all equal to MISSING_VALUE or EMPTY_VALUE. This is useful
to force information to be printed if it is specifically requested
through a query or through specifying the filters manually.

Test Description:
Run unit tests to ensure funcationality remained the same. Test various
commands which use the --filter functionality and ensure that they
retain only empty/missing columns when --show-empty or --show-missing
are not given.