CRAYSAT-189: Add option to only report specified columns
This change adds an option to the Report class which allows users to
specify certain columns which should be displayed in the output of
commands. Columns which are present in the new `display_headings`
parameter but not in the `headings` parameter are ignored, and a warning
is logged. Similarly, when duplicate headings are passed to
`display_parameters`, the duplicates are ignored and the first instance
is used, and a warning is logged.

This change also adds a `--fields` argument to all subcommands using the
format options group. The argument passed to the `--fields` option is
passed into the Report class' `display_headings` argument in order to
limit the output columns. Column limiting is performed after filtering,
so filters can be applied to columns that are not displayed, but are
still present in the output. (This behavior is similar to SQL's "SELECT
... FROM ... WHERE ..." statements.)

Finally, this change also adds unit tests to test the new functionality
added to the Report class to support the `--fields` option.

Test Description:
Ensure existing and new unit tests pass. Manually test the behavior of
the `--fields` option with all affected subcommands to ensure it works
as expected.