CRAYSAT-634: Add improved error messages for filtering
This change refers users to the manual pages related to filtering when
an error is encountered with filter syntax. Since bash can interpret
quotes as it sees fit, SAT may not have enough information to go on
regarding syntax errors, and so referring the users to read the man page
is useful as we can't always detect when bash is changing the input from
what the user typed verbatim.