CRAYSAT-949: Add option to exclude NCNs from bootsys stages
Add option to exclude the NCNs specified by the user from the actions
that are performed in the platform-services and ncn-power stages of the
boot and shutdown actions in `sat bootsys`.

Also added a prompt to the ncn-power stage for the user to confirm the
NCNs have been properly determined for consistency with the
platform-services stage.

Also added a `BeginEndLogger` context manager to the `do_power_on_ncns`
function for consistency with the `do_power_off_ncns` function.

Test Description:
pycodestyle passes. Unit tests pass.

Tested on a vshasta cluster. Tested with and without `--excluded-ncns`
option.

Built man pages and inspected sat-bootsys.8 in `man` viewer.