CRAYSAT-949: Reduce scale of test of remove_constant_values
Reduce the scale of the test of `remove_constant_values` in
`sat.filtering` module. The algorithm is O(m*n) where m is the number of
keys in each dict, and n is the number of rows. Since it scales
linearly with the number of rows, we don't really need to test with a
huge number of rows to verify that its performance is acceptable. Reduce
the scale to to one fifth of its original scale. The call to
`remove_constant_values` now takes about 0.07 seconds when I ran it with
`nosetests` locally.

Most of the time is actually spent constructing the input value. Get rid
of the random calls because those add significant time.

Test Description:
Tested with `nosetests` configured to run only this test. Modified
`remove_constant_values` to take longer, and verified that the test
failed and gave a message about the time being exceeded.