CRAYSAT-1027: Fix issue when sensor metrics are null
This change fixes a small bug in `sat sensors`. In some instances, the
metric sensor result can be set to `None` due to a null value being
returned from the SMA API. This sometimes caused an issue with
`make_raw_table()` which would cause a traceback to be printed.  This
change also adds a test which checks for this edge case.

Test Description:
Reproducing this bug on a running system is pretty inconsistent, so
testing is done automatically with the unit test suite.