CRAYSAT-988: Add SDU checker to bootsys session checks
This change adds a new session checker for SDU to the session-checks
stage of `sat bootsys shutdown`. This ensures that no SDU sessions are
running while a bootsys operation is in progress. If SDU is not set up
on the system, or the SDU container is not running, this session check
report a warning, but allow the user to continue. Otherwise, this check
will report if an active SDU dump is in progress by searching for an SDU
process on all manager NCNs.

This change also refactors the `ServiceActivityCheck` class slightly,
making it a subclass of the more general `ActivityCheck` class.
ServiceActivityChecks are intended to interact with services accessible
with the `cray` cli, while ActivityChecks may check for any ongoing
activity on the system.

Test Description:
Automated tests were added to cover the added functionality. The
session-checks stage was also tested manually on gamora while SDU
sessions were running, as well as when they were not. In addition,
error cases were tested (e.g. SDU not installed, container not running,
etc.)