CRAYSAT-1048: Warn on hosts without known host key in bootsys
Some locations in the `sat bootsys` code issued a warning when a host with
an unknown host key was encountered, and locations rejected the attempt
to connect to the host. Consistently use the `WarningPolicy` instead of
the the default `RejectPolicy` throughout the `sat bootsys` code.
Create a simple function that creates an `SSHClient`, loads the system
host keys, and sets the missing host key policy to `WarningPolicy`.

Add unit tests for the new function and modify existing unit tests to
mock out this `get_ssh_client` function instead of the `SSHClient` class
itself.

Test Description:
Unit tests pass. pycodestyle reports no violations.

Tested on vshasta by removing the entry for ncn-m001 from
`~/.ssh/known_hosts` and then ran through the following sequence of
commands:

* `sat bootsys shutdown --stage platform-services`
* `sat bootsys shutdown --stage ncn-power`
* `sat bootsys boot --stage ncn-power`
* `sat bootsys boot --stage platform-services`

A warning was printed each time a connection was made to ncn-m001, but
each command completed successfully.

Then added ncn-m001 back to `~/.ssh/known_hosts` and ran through the
same test and verified that no warnings were printed and commands
completed successfully.