CRAYSAT-1044: Fix sat hwinv --fields behavior
This change is a one-line amendment to a previous fix for the behavior
of the --fields option with sat hwinv listings. The option for
`--fields` was always used, even when a more component-specific option
was specified simultaneously. This change fixes the specificity of field
filtering options.

Test Description:
Test manually to verify the correct fields are used on thanos.