CRAYSAT-998: Restrict sat diag to Rosetta switches
This change adds a restriction requiring all xnames given as input t sat
diag be marked as type `RouterBMC` in HSM so that diags may only be run
on Rosetta switches. Components of other types are not supported. If an
xname of a component is given which does not correspond to a Rosetta
switch, an error is logged and sat will exit without running
diagnostics.

An additional option, `--no-hsm-check`, was also added to bypass
checking if the given xnames are supported Rosetta switches. This can be
useful in situations when HSM is unavailable.

Test Description:
Manually run sat diag with xnames of Rosetta switches and
non-Rosetta-switch components. Test behavior works as expected with and
without --interactive. Test behavior of `--no-hsm-check` to ensure the
check can be bypassed if needed.