CRAYSAT-871: Refactor multiple xname firmware fetching
This change refactors the code path for fetching firmwares for multiple
xnames. These changes improve the performance characteristics of
firmware queries with many xnames; instead of the walltime required for
the queries increasing linearly with the number of xnames, the amount of
time remains roughly constant. Some dead code was also removed, and
corresponding unit tests were also removed.

Test Description:
Test with and without a pre-existing snapshot, and with and without the
`--xname` option specified. Ensure performance remains O(1) as xnames
are added.