CRAYSAT-328: Fix log file permission error.
More gracefully handle the case when the user does not have permissions
to write to the default log directory or to the default log file.

Change errors to warnings when logging to file cannot be set up because
these are not fatal errors. The command will continue on, but it just
won't log to a file.

Test Description:
Tested locally in five cases:
* User has insufficient permissions to create a new log file
* User has insufficient permissions to write to existing log file
* Log file path already exists but is a directory
* Log file does not exist and user has permissions to create it
* Log file does exist and user has permissions to log to it