CRAYSAT-1121: Wait for ComputeModules in cabinet-power stage
This change causes `sat bootsys boot --stage cabinet-power` to wait for
ComputeModules instead of NodeBMCs to be powered on. This change fixes a
bug encountered during liquid-cooled cabinet bringup as CAPMC does not
provide power status of BMCs, so we have to check the status of
ComputeModules instead. Waiting for this state works as expected.

Test Description:
Run a "dry-run" on shandy while the machine is already booted.
Successfully wait for the hms-discovery cronjob to be scheduled, and
then wait for the nodes to reach the "on" state. (shandy had 2 compute
modules marked Empty, though this was for hardware testing reasons, and
the rest of the modules that were on were waited for as expected.)