CRAYSAT-1057: Show node aliases in `sat status`
This change adds a new feature to `sat status` which retrieves node
aliases (i.e. hostnames) from SLS to display in the status output. This
allows users to more quickly get a list of hostnames of nodes available
on the system.

This change also add unit tests for the new functionality, and updates
the man page to reflect new output.

Test Description:
Manually run `sat status` on shandy and ensure that output matches the
state of the system. Run unit tests.