CRAYSAT-552: Add WaitingFailure mechanism for Waiters
This change adds a new Exception type for use with the Waiter class,
`WaitingFailure`. This exception can be raised by the
`Waiter.has_completed()` or  `GroupWaiter.member_has_completed()`
methods. This exception being raised indicates that the given condition
has not been reached, and that the condition cannot be reached due to
the state of the system. When raised, this exception will cause the
Waiter to stop waiting for the condition, or for the GroupWaiter to stop
waiting for the given member to reach the condition. Appropriate error
messages are logged based on the information given in the WaitingFailure
exception.

Test Description:
Ensured unit tests continue to pass.