CRAYSAT-920: Specify all components to power off for CAPMC
This change gets all the components to be powered off (Chassis,
ComputeModules, and RouterModules) from HSM and passes them individually
to CAPMC instead of passing `"recursive": "true"` in the request. This
works around a bug in CAPMC which causes the whole request to fail when
recursively powering off components which have disabled sub-components
underneath them in their trees.

When querying the list of components to use in a power operation, we
no longer pass omit_empty=False per discussion with Michael Jendrysik
and Manny Alvarez. Components should not be marked Empty during power on
or power off; they should only be marked Empty when they are physically
removed from the system.

Test Description:
Disable a node on rocket and power off the system per the manual. Ensure
the shutdown process proceeds correctly.