CRAYSAT-1083: Rename create_cfs_configurations
Rename create_cfs_configurations to just create_configurations to make
the concept of creating configurations more generic at the top level.
This is being done for symmetry with a new create_images function that
is being written to create images. This should more easily allow future
extensibility if we support different kinds of configurations and/or
images.

Test Description:
Unit tests pass.