CRAYSAT-1198: Implement image customization in `sat bootprep`
Implement the customization of images using CFS sessions in `sat
bootprep`.

Log info about the containers within the pod for the CFS session as they
complete, so that the admin can see when each container for a CFS
session completes and whether it is successful or failed.

This includes a reorganization that puts the classes that represent the
input data from the input file into their own subpackage, `input`,
within the `sat.cli.bootprep` subpackage. The unit tests are reorganized
accordingly as well.

Test Description:
Unit tests pass. pycodestyle passes.

Tested with minimal single image and configuration on baldar. Needs more
testing of failure cases and multiple images at once.