CRAYSAT-1135: Show which version is active in showrev
This commit modifies the `sat showrev` command so that if an 'active'
field is present and set to "true" for a given product, it will be
displayed this way in the output with an 'active' column.

A unit test is added for this behavior, and the unit tests for getting
product revision information are changed slightly to reduce duplication
of mock data. As a part of this, a couple missing docstrings are added
to test case classes in test_products.py.

Finally, the man page for SAT is updated to include a reference to
showing the active version, and the change log is updated as well.

Test Description: I ran this version of SAT on an internal system
where some product versions were set to "active".

Test Description:
I installed this version of SAT in a virtualenv on an internal system
and ran `sat showrev`, where there were both active/inactive versions
of SAT and versions of other products for which there was no specified
active version.