CRAYSAT-1198: Fix typos and sonarlint warnings in unit tests
Fix a typo that occurs in two places in the unit tests for the git and
product configuration layer classes.

Resolve a sonarlint warning about a string literal that is duplicated 8
times by defining a parent class for the individual git and product
layer test classes. This parent class provides a method to patch the
value of `InputConfigurationLayer.resolve_branches`.

Fix a mistake in the name of the `TestProductInputConfigurationLayer`
class and in its docstring.

Test Description:
Unit tests pass. pycodestyle passes.