CRAYSAT-1175: Add a subpackage for querying software inventory
Add a subpackage to sat that provides an abstraction for querying for
software inventory from the cray-product-catalog Kubernetes config map.
Most of this code is borrowed from the `shasta-install-utility-common`
python package.

Add support to the `software_inventory` subpackage to get the latest
version of an installed product from the product catalog. This is needed
for the `ProductInputConfigurationLayer` functionality in `sat
bootprep`.

Add unit tests for the software_inventory subpackage. These tests are
adapted from the similar unit tests in the shasta-install-utility-common
library from which this subpackage originated.

Note that this adds a dependency on the `cray-product-catalog` python
package which currently contains the product catalog schema and
validation function.

Eventually, the goal is to migrate this new `software_inventory`
subpackage into its own top-level package. At that time, it would also
make sense to add the schema validation code to this new package. The
new package would then be used uniformly across all of the following
components that use it for accessing product catalog data and validating
it against the schema:

* sat
* shasta-install-utility-common
* cray-product-catalog

Test Description:
Unit tests pass.