CRAYSAT-1293: Substitute API hostname in clone URL (#434)
CRAYSAT-1293: Substitute API hostname in clone URL

This commit changes the clone URL parameter for CFS sessions that
are created from a given product-branch or product-version. Previously
this value was pulled directly from the product data in the product
catalog, but it was not possible for CFS to clone using that URL, due
to this being an external address on the customer access network (CAN).

This commit substitutes the CAN address for the local API gateway
hostname as configured in the SAT configuration file.

This commit also updates the unit tests to test for this behavior.

Test Description:

I installed this branch of sat in a virtualenv on an internal system,
then I ran `sat bootprep --dry-run --save-files --input-file FILE` three
times, with the following input files:

* An input file with one configuration, whose only layer specifies a
  product and version.
* An input file with one configuration, whose only layer specifies a
  product and VCS branch.
* An input file with one configuration, whose only layer specifies an
  explicit git URL and branch.

For the first two inputs, I checked the resulting cfs-config-CONFIG.json
files that were saved by using `--save-files`, checking that the clone
URLs were substituted. For the final input, I checked that the resulting
clone URL was NOT changed, as this was explicitly set to a certain URL.