CRAYSAT-1175: Ensure the output directory exists if necessary
Ensure the output directory exists if necessary, i.e. when the action is
`run` and the `--save-files` option is specified or when the action is
`generate-docs` or `generate-example`.

Test Description:
Tested by running all of the following commands in the case when the
output directory already exists, exists as a file, and does not exist:

* `sat bootprep run --output-dir output --save-files --dry-run configurations.yaml`
* `sat bootprep generate-docs --output-dir output`
* `sat bootprep generate-example --output-dir output`