CRAYSAT-1268: Enhance get_val_by_path
This commit enhances get_val_by_path to handle when a dictionary value
is a NoneType rather than a dictionary. This makes it possible to use
IMSClient.rename_image to rename an image that has no manifest file
(i.e. an empty image). Add a unit test for this new behavior.