CRAYSAT-1268: Permanently delete an image in rename_image
Renaming an image is done by first copying an image and then
deleting the original image. However, this delete simply
moves the original images into /deleted/images/. This change
updates the delete_image method of IMSClient so that an
image can be permanently deleted (i.e., does not end up
in /deleted/images/) when the parameter permanent=True is given.
The call to delete_image in rename_image is updated to use
permanent=True.

Test Description:

* Created an IMSClient and ran delete_image with and without
  permanent=True, checking that the deleted image was absent
  from /deleted/images only when using permanent=True
* To test this commit and the previous one, I created an empty
  image, created an IMSClient and ran rename_image on it,
  checking that the original image was not present in the
  list of deleted images.