CRAYSAT-1264: Add suggestions of k8s logs to view on bootprep failure
This change improves the implementation of
CFSImageConfigurationSession.get_suggested_debug_commands() to suggest
viewing logs for only the most relevant containers which failed in a CFS
session. This change also adds a call to the above function to the
bootprep impelmentation so that suggested commands are logged on
failure.

Test Description:
Test on internal system with specially-constructed bootprep input files
which contained CFS configurations with non-existent git repo URLs in
various positions in the configuration layer stack. Test with
specially-constructed configuration ansible playbooks which fail in
various points in the configuration layer stack and test that the proper
kubectl log command is logged.