CRAYSAT-1266: Improve CFSImageConfigurationSession.update_status
Improve the robustness of the `update_status` method of
`CFSImageConfigurationSession` by only checking for the pod status when
the K8s job has been scheduled according to the session status reported
by CFS. Additionally, in case the job exists, but the pod does not yet
exist, just log a message indicating that the session is waiting for its
pod to be created instead of failing.

Test Description:
Tested by running `sat bootprep` to create and configure a barebones
image on lemondrop.