CRAYSAT-1292: Allow specifying a commit hash in product layer
This change allows users to specify a commit hash from a configuration
repository relating to a specific product. Previously a product layer
could support only a product name and version, as well as an optional
branch name.

An example product configuration layer with a commit hash might look as
follows:

```yaml
  - name: uan-integration-x.y.z
    playbook: site.yml
    product:
      name: uan
      commit: c07f317c4127d8667a4bd6c08d48e716b1d47da1
```

This change also adds unit tests for the new feature.

Test Description:
Run unit tests. Test on internal system using test configurations which
create small test configurations with layers which specify commits, and
ensure that the configurations were created properly. Test with
configurations which use products and branches as before to test for
regressions.