CRAYSAT-1324: Split `sat status` by component type
This change splits the output from `sat status` into multiple report
tables in the event that multiple types of components are specified.
This makes it a bit easier to differentiate the statuses of different
types of components, and prevents many columns from showing "MISSING"
values for certain components.

Additionally, the "Type" column has been removed and replaced with a
report header. The header is only shown if multiple reports for
different types are shown.

The man pages were updated to reflect this change.

Test Description:
Ran unit tests, tested on an internal system in a virtualenv by
specifying multiple types specifically, specifying `--types all`, and
testing the default behavior.

Co-authored-by: Ryan Haasken <ryan.haasken@hpe.com>