CRAYSAT-1184: Improve StatusModule unit test isolation
Improve isolation between unit tests so that no unit tests are leaving
cruft in the real `StatusModule.modules` class attribute. That attribute
will be replaced with an empty list at the beginning of each unit test
and restored to the real `StatusModule.modules` at the end of the unit
test.

Add a unit test that does a sanity check that all unit tests are well
isolated from each other and not leaving cruft in
`StatusModule.modules`.

Test Description:
All unit tests pass.