CRAYSAT-1184: Improve sanity-check unit test
This changes the sanity-check unit tests which ensures that subclassing
`StatusModule` in the unit tests does not actually add them to the
"real" `StatusModule.modules` list. The length of the list to check
against is now introspectively computed from the contents of the
`sat.cli.status.status_module` module using the `inspect` module.

Test Description:
Run unit tests.