CRAYSAT-1331: Add --delete-ims-jobs option
This change adds a `--delete-ims-jobs` option to `sat bootprep` which
causes IMS image creation jobs to be deleted after completion. The
default behavior was also changed to not delete image creation jobs; it
is now opt-in with the new command line flag.

Test Description:
Run `sat bootprep` on an internal system with small barebones images,
testing both with and without `--delete-ims-jobs`.