CRAYSAT-1365: Make `sat bootprep` require an action argument
The action should be required for `sat bootprep` calls. Set `required`
to `True` on the `actions_subparsers` this way instead of passing it as
a keyword argument in `add_subparsers` because that keyword arg is not
added until Python 3.7, so we might as well do it in the way that is
more compatible with older Python 3 versions.

Test Description:
Tested that running `sat bootprep` without arguments raised an error.
Tested this approach with Python 3.6.9 and Python 3.8.8 in local pyenv
virtualenvs.