CRAYSAT-1393: Handle when FAS returns empty strings for 'error'
This commit handles the case when FAS returns valid responses for
an xname but either the xname or the "targets" under the xname
have an "error" key but the value of it is set to an empty string.

Previously, when such a field existed for an xname, we would skip
adding the xname's targets to the report. Additionally, we would log
an error which would amount to "Error: <empty>". Now, we skip logging
an error and still attempt to add all the targets to the report
(provided they have the required fields).

This commit also changes how "empty string" errors are handled when
they are returned for an individual target under an xname. Previously,
an "empty string" error would result in an error being logged, and
now they will not.

This commit adds unit tests for the new behavior.

Test Description:
* Pushed this branch and built image in Jenkins
* Downloaded image onto hela-ncn-m001
* Ran 'sat firmware' and confirmed expected output