CRAYSAT-1401: Rewrite --bos-template man page description
This change rewrites the description of the `--bos-template` option to
the `sat status` subcommand, specifying that the list of nodes in the
given BOS may not match that of the booted session template if changes
have been made to roles, groups, or the node list.

Test Description:
Build man pages and inspect for correctness.