CRAYSAT-1420: Fix missing `sat hwhist` man page
The man page for `sat hwhist` was not being built by the `Makefile`
because it was not listed as a prerequisite of the `.all` target. Fix
this issue and prevent it from happening again in the future when we add
a new command by changing the `Makefile` to list all files ending in the
suffix `.8.rst` and then strip off the `.rst` and use that as the
prerequisite list of the `all` target.

Update developer docs accordingly.

Test Description:
Tested by running `make clean` followed by `make all`. Confirmed that
all man page files ending in `.8.rst` were built to corresponding
`.8` man page files. Examined a couple to make sure they looked correct.
Modified a single `.8.rst` file and checked that `make all` just rebuilt
that one.