CRAYSAT-1426: Fix issue with `sat status` traceback
This change fixes a regression in `sat status` which caused a traceback
when an xname was missing from SLS. In this situation, the `Aliases`
field wasn't able to be populated in that row, and a KeyError was thown.
This change fixes the bug by initially filling each non-primary column
of each row in as 'MISSING' before retrieving data from the appropriate
service and filling in the columns with that information.

This change also adds a regression test to check for this condition and
proper handling.

Test Description:
Run unit tests. Run on internal system with xname missing from SLS and
check that the xname's `Aliases` field is listed as `MISSING`.