CRAYSAT-1361: Remove docker from requirements files
We no longer depend on the `docker` python package in `sat`. We used to
use it for the `--docker` option of `sat showrev` to list local docker
images, but that usage has been removed.

This removal was done using the following procedure:

1. Created a new python virtual environment with python 3.9.7, which
   matches the version in our current version of Alpine Linux.
2. Activated the virtual environment.
3. Removed `docker` from `requirements.txt` and `requirements-dev.txt`.
4. Ran `pip install -r requirements.txt -c requirements.lock.txt`
5. Ran `pip freeze > requirements.lock.txt`
6. Ran `pip install -r requirements-dev.txt -c requirements-dev.lock.txt`
7. Ran `pip freeze > requirements-dev.lock.txt`

Test Description:
Will sanity test the resulting image by downloading it to a system and
running some `sat` commands.