CRAYSAT-79: Remove warning when session is empty string
This change removes a warning message printed when a given component's
session is the empty string. This can happen when a node hasn't been
booted by a BOS session, and is not necessarily a problem. A warning is
still printed if the "session" key is not present.

Test Description:
Test on internal system and ensure that no warnings are printed for
components which have `.session == ""`.