CRAYSAT-1409: Avoid unnecessary check on existing items
When no items are specified in a property of the input file (currently
this just pertains to `session_templates`), skip the check on existing
items. Otherwise, we are performing unnecessary API calls to get the
list of existing items (e.g. API calls to BOS to get session templates).

Test Description:
Tested that I could run `sat bootprep run` against a bootprep input file
that contained `session_templates: []` in a virtualenv on my macbook.