CRAYSAT-1409: Allow `sat bootprep` to find image/recipe from product
Update the `sat bootprep` input file schema to allow the user to specify
images which are created from images or recipes provided by a product.
Update the `sat bootprep` code to query the `cray-product-catalog` K8s
ConfigMap to find the image/recipe provided by a product.

Note that this does not handle products which deliver multiple recipes
or images yet. If an image or recipe is requested from a product, and it
delivers more than one image or recipe, an error will be returned to the
user.

Due to the bootprep schema update, validation unit tests had to be
updated to find errors at their new indentation levels since the
top-level error for an invalid image will be of the form "Not valid
under any of the given schemas", and the more specific error messages
will be indented an additional level.

Refactor the `render_template` function which was only used by the
`sat.cli.bootprep.input.configuration` module to a shared location and
rename to `jinja_rendered` to sound more like a decorator. This
decorator is now used in CFS configurations (and layers) and in images
from the input instance.

Also use a `SandboxedEnvironment` when rendering Jinja2 templates to
avoid possible security vulnerabilities that could arise from rendering
untrusted Jinja2 template provided by the user. The risk of this sort of
attack using `sat bootprep` is low though given that SAT is a local
command-line interface which runs as the user invoking it.

Test Description:
Unit tests and pycodestyle pass.

Built docker image and using that docker image on vale, tested the
following scenarios:

* Creation of CFS configurations which specified product version
  variables.
* Creation of image from a base recipe provided by the cos product.
* Customization of an image from a base image provided by the cos
  product.
* Mocked out a product catalog entry to have no recipes/images, and
  specified the image and recipe from that product. Error resulted.
* Mocked out a product catalog entry to have multiple recipes/images,
  and specified the recipe/image from that product. Error resulted.
* Tried a malicious Jinja template of the form `"".__class__.__mro__`
  to validate that the `SandboxedEnvironment` provides some protection
  against malicious Jinja2 templates supplied by the user.