CRAYSAT-1446: Support swapping Hill blades
This change supports swapping Hill hardware in the same manner as
Mountain hardware. Previously this would not work as `sat swap blade`
explicitly checked only for "River" and "Mountain", and "Hill" was
excluded arbitrarily.

Test Description:
Use `sat swap blade -a disable` to swap out two blades on a Hill system.
Physically swap the blades. Use `sat swap blade -a enable` to swap back
in the blades, and boot to finish the test. Run unit tests.