CRAYSAT-1411: Implement image and recipe name substitution
Add the ability to use the name of the base image or recipe that was
used to construct an image as a variable substitution in the name
property of an image.

Also add general variable substitution (e.g. product version variable
substitution) to the conifguration property for images and to the name,
image, and configuration properties of session templates.

Test Description:
Not tested yet.