CRAYSAT-1410: Allow session templates to refer to images by ref_name
Add functionality to `sat bootprep` to allow session templates in the
input file to refer to images by their `ref_name` value. This is a
usability improvement to make it easier to refer to an image which has a
name that uses its base image or recipe name in its name.

Deprecate the old way of specifying the IMS image name or ID as a string
in the session template. Add the ability to specify an object under the
`image` property that has either an `ims` property or an `image_ref`
property. The value of the `ims` property is a dict with either a `name`
or an `id` property. The value of the `image_ref` property is a string
that refers to an image in the input file by the value of its `ref_name`
property.

Test Description:
Tested on vale with a bootprep input file that creates session templates
which specify their images using `image.ims.id`, `image.ims.name`, or
`image.image_ref`.

Also tested creating a session template with `image` set to a string
containing either the `id` or the `name` of an image.

Tested the following scenarios in dry-run mode, which were initially
broken and then fixed by Eli's suggest changes:

* A configuration that builds an image, then creates a session template
  that uses that image, using the "old"-style format of referring
  directly to the image by name.
* A configuration that also builds an image and session template with
  that image, but using the "new"-style format of referring to the image
  using `ref_name` and `image_ref`.