CRAYSAT-1502: Update the name of the VCS repo
The name of the repository created to contain recipe information was
changed to hpc-shasta-software-recipe, so change the name used by the
`sat bootprep` code to match.

Test Description:
Minor change. Will be tested with integration of SAT with new
hpc-shasta-software-recipe release distribution prior to merge of
feature branch to integration.