CRAYSAT-1411: Default to using the latest software recipe version
When loading variables to use as context in Jinja2 template rendering of
`sat bootprep` input file fields, default to looking for the latest
version of the HPC software recipe. Also change the behavior when the
variables can't be loaded from the HPC software recipe to log a warning
instead of fail with an error. It made sense to fail with an error when
a recipe version was explicitly requested on the command-line, but now
that we default to look for the latest, it makes sense to log a warning
and continue in case the input file does not need those variables or
they are defined with `--vars-file` or `--vars`.

If the input file does need those variables, Jinja2 rendering will fail
and an error message will be printed about the missing variables.

Test Description:
Tested on vale with docker image from Jenkins build. Tested the
following scenarios:

* No `--recipe-version` specified.
* An unknown `--recipe-version` specified.
* An alternative, non-latest `--recipe-version` specified.