CRAYSAT-1536: Remove image dependencies when skipped
When an image depends on another image being created by the input file,
and that latter image already exists and re-creation of that image is
skipped, remove that latter image as a dependency of the former image.
This ensures that all images in the input file will be created and exist
by the time we get to the point of creating BOS session templates, which
may reference those images.

Add a `skip` attribute to `BaseInputImage` and set it to true if the
user requests to skip the re-creation of an existing image. Use this
attribute in the computation of the `ims_base` property of the
`DependentInputImage`, so that we can find the ID of the base image when
that dependency image already existed and was not created by this
invocation of `sat bootprep`.

Test Description:
Created a bootprep input file that defines three images. The first
image is built from the recipe provided by the COS product. The second
image uses the first image as a base and customizes it for computes. The
third image uses the first image as a base as well but customizes it
differently for UAN. The configurations used were not real CN/UAN
configurations, but rather simple CFS configurations with a debug
playbook. The bootprep file also contained two session templates, one
that referenced the CN image and one that referenced the UAN image.

Ran `sat bootprep run` against this input file with the CFS
configuration designed to fail. Ran `sat bootprep run` against this
input file again with the CFS configuration designed to succeed.
Answered the prompt about the existing image with "skip", so the base
image would not be created again. Confirmed that the other two images
were still created, and creation of the BOS session templates succeeded.